"Add metre to mile conversion function and use it to display store distance in FindStore component"
This commit adds a new function `metresToMiles` to the `parse-util.ts` file which converts meters to miles. This function is then used in the `FindStore.tsx` component to convert the `DistanceInMetres` property of each store to miles and display it in the UI. Additionally, the code now checks if the distance is less than 0.1 miles and displays it as "less than a mile" instead of "0.01 miles".