"Add optional featureToggle prop to LayoutError and ProductLayout components"
This commit adds an optional `featureToggle` prop to the `LayoutError` and `ProductLayout` components, which are functional components that receive various props related to the layout of a webpage. The `featureToggle` prop is of type `any` and is added to the existing props interface for both components. Additionally, the `featureToggle` prop is passed to the `MainNav2Logged` component in the render method of both `LayoutError` and `ProductLayout`.