Redirect to review page for partial payment failure
When a partial payment fails, redirects the user back to the review step of checkout instead of the generic payment failed page. This allows the user to correct any issues and retry the payment more easily. The cart is checked for `isPartialPayment` to determine the correct redirect URL. Also, the `getCart` method is now being pulled from the `cartHandler` hook.